requirements.txt
streamlit>=1.30.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.10.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
python-docx>=1.0.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
tiktoken>=0.7.0
